SmartDrive-Dashboard/requirements.txt
flask
opencv-python
ultralytics
picamera2
RPi.GPIO
torch
